### 2021-06-18 #3382 @whedon
[PRE REVIEW]: Epimargin: A Toolkit for Epidemiological Estimation, Prediction, and Policy Evaluation
**Submitting author:** @satejsoman (<a href="http://orcid.org/0000-0001-8450-7025">Satej Soman</a>)
**Repository:** <a href="https://github.com/COVID-IWG/epimargin" target ="_blank">https://github.com/COVID-IWG/epimargin</a>
**Version:** v0.5.2
