+++ paper/paper.bib
@@ -50,7 +50,8 @@
   number={6532},
   pages={916--921},
   year={2021},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1101/2020.09.08.20190629}
 }
 
 @article{bettencourt2020systems,
@@ -58,14 +59,16 @@
   author={Bettencourt, Luis and Soman, Satej},
   journal={Mansueto Institute for Urban Innovation Research Paper},
   number={26},
-  year={2020}
+  year={2020},
+  doi={10.2139/ssrn.3748704}
 }
 
 @techreport{malani2020adaptive,
   title={Adaptive control of COVID-19 outbreaks in india: Local, gradual, and trigger-based exit paths from lockdown},
   author={Malani, Anup and Soman, Satej and Asher, Sam and Novosad, Paul and Imbert, Clement and Tandel, Vaidehi and Agarwal, Anish and Alomar, Abdullah and Sarker, Arnab and Shah, Devavrat and others},
   year={2020},
-  institution={National Bureau of Economic Research}
+  institution={National Bureau of Economic Research},
+  doi={10.3386/w27532}
 }
 
 @techreport{southasiavaccinates,
