### 2021-07-02 #3382 @Nikoleta-v3
:wave: @edrictam & @ayush9pandey & @wxwx1993 & @BrandonEdwards, would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "Epimargin: A Toolkit for Epidemiological Estimation, Prediction, and Policy Evaluation". You can find more information at the top of this Github issue 🆙

Please let me know if you're available :smile: Thank you!
