### 2021-07-09 #3382 @Nikoleta-v3
Thank you @wxwx1993 😄

The official reviewing guidelines for JOSS are available here: https://joss.readthedocs.io/en/latest/review_checklist.html

The reviews are checklist-driven. As you can see the current issue is labeled a `pre review` issue. Once we find a second reviewer this issue will be closed, and we will open the `review` issue. There you will find a checklist that you need to complete as a reviewer. The checklist is there mainly to help/guide you.

If you have any issues you want to raise regarding the submission feel free to open issues in the original repository and/or comment on the `review` issue. Hope this information helps!

Now regarding deadlines. Though there is not a hard deadline I would appreciate receiving your review within a reasonable time.  I could set a 2 weeks reminder for you once the review process begins.
