### 2021-07-09 #3382 @Nikoleta-v3
👋 @dilawar & @nhtran93 & @krother & @Genomematt, would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "Epimargin: A Toolkit for Epidemiological Estimation, Prediction, and Policy Evaluation". You can find more information at the top of this Github issue 🆙
