### 2021-07-09 COVID-IWG/epimargin#115 @dilawar
Invalid dependency version `zipp==4.1.0`
https://github.com/openjournals/joss-reviews/issues/3464

https://pypi.org/project/zipp/#history does not have `zip==4.1.0` as specified in the requirements.txt file. I am on debian buster with python3.7.

Here is a summary of `pip install -r requirements.txt`.

```
Collecting xlrd==1.2.0
Downloading xlrd-1.2.0-py2.py3-none-any.whl (103 kB)
ERROR: Could not find a version that satisfies the requirement zipp==4.1.0 (from versions: 0.1.0, 0.2.0, 0.2.1, 0.3.0, 0.3.1, 0.3.2, 0.3.3, 0.4.0, 0.5.0, 0.5.1, 0.5.2, 0.6.0, 1.0.0, 1.1.0, 1.1.1, 1.2.0, 2.0.0, 2.0.1, 2.1.0, 2.2.0, 2.2.1, 3.0.0, 3.1.0, 3.2.0, 3.3.0, 3.3.1, 3.3.2, 3.4.0, 3.4.1, 3.4.2, 3.5.0)
ERROR: No matching distribution found for zipp==4.1.0
```
