### 2021-07-09 #3464 @dilawar
Below is a partial review. It includes major issues I found during the installation and testing the software. I am (doker images of Linux: cenos8, debian:buster, and ArchLinux). The `setup.py` says that software is `OS Independent` and supports `>=python3.6`.  I am mostly using python3.9 and python3.8.

In my fork, I've also added a Github actions based CI and a few test cases.

Yes. `pip3 install apimargin` succeeded on a Linux machine: it downloaded a platform-independent wheel. The installation from the source failed on my machine because of pinned version in `requirements.txt`. I think that `requirements.txt` file can be removed since `setup.py` files also enumerate the installation dependencies.

After installation (from pypi), however, I ran into permission issues at the very first run.

from itertools import cycle

import epimargin.plots as plt
import pandas as pd
from epimargin.utils import setup

(data, figs) = setup() # creates convenient directories
plt.set_theme("minimal")
```

resulted in,

```sh
<ipython-input-1-710a13386d8f> in <module>
      6 from epimargin.utils import setup
      7
----> 8 (data, figs) = setup() # creates convenient directories
      9 plt.set_theme("minimal")

~/Work/FORKES/epimargin/epimargin/utils.py in setup(**kwargs)
     42     logging.basicConfig(**kwargs)
     43     logging.getLogger('flat_table').addFilter(lambda _: 0)
---> 44     return (mkdir(root / "data"), mkdir(root / "figs"))
     45
     46 def fillna(array):

~/Work/FORKES/epimargin/epimargin/utils.py in mkdir(p, exist_ok)
     30
     31 def mkdir(p: Path, exist_ok: bool = True) -> Path:
---> 32     p.mkdir(exist_ok = exist_ok)
     33     return p
     34

/usr/lib/python3.9/pathlib.py in mkdir(self, mode, parents, exist_ok)
   1311         """
-> 1313             self._accessor.mkdir(self, mode)
   1314         except FileNotFoundError:
   1315             if not parents or self.parent == self:

PermissionError: [Errno 13] Permission denied: '/usr/bin/data'

```

I solved it by modifying the `cwd` function (diff: https://github.com/dilawar/epimargin/commit/98535fb8921ff1b58ddff8324f83f469bc6e6671#diff-eb56eed78c526134068ae48c66f6eaaf70109da1d0fc031cc83cdd3e78285fff). May authors can review and use this commit. Open issue on repo https://github.com/COVID-IWG/epimargin/issues/116

Yes.

    1. `setup.py` has list of dependencies which `pip` handles well.
    2.  A `requirements.txt` file is also given with pinned version of dependencies. User is expected use this file: `pip3 install -r requiements.txt` before `pip3 install -e .` but it did not succeed on Ubuntu/Debian and CentOS:8, For example, I did not find `zip=4.1.0` (https://github.com/COVID-IWG/epimargin/issues/115) anywhere in pypi. I had to unpin few dependencies so `pip` can resolve an install-able version. Probably there is no need of `requirements.txt` file if dependencies are handled in `setup.py`.

Yes. A simple example is given in the README file. I was able to reproduce the figure.

And a lot of examples are available in `studies` folder. But instructions to run them are lacking. Ideally, each folder should have a pipeline based on`Makefile`, `bash` scripts etc.. Or authors should mention the steps to reproduce the results in README file in the folder itself. A few folders do not pass the basic lint tests (issue #117).

### Some failed cases on CentOS-8

#### 1

```sh
[centos@ip-172-26-6-166 karnataka (joss_lint_errors)]$ python3 age_rt.py
WARNING (theano.tensor.blas): Using NumPy C-API based implementation for BLAS functions.
  File "/home/centos/Work/FORKES/epimargin/studies/karnataka/age_rt.py", line 3, in <module>
    from epimargin.estimators import analytical_MPVS
  File "/home/centos/Work/FORKES/epimargin/epimargin/estimators.py", line 7, in <module>
    import pymc3 as pm
  File "/home/centos/.local/lib/python3.9/site-packages/pymc3-3.11.2-py3.9.egg/pymc3/__init__.py", line 34, in <module>
    if not semver.match(theano.__version__, ">=1.1.2"):
AttributeError: module 'semver' has no attribute 'match'
```

Authors should rerun their examples with python-3.9 or limit the python version in `setup.py` file.

- [ ] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

No. There are no automated tests. If there are some tests deep inside the codebase, I can't find a them via `pytest` etc. Authors should indicate somewhere in README.md if there is a test suite and how to run it. I've created one during my review https://github.com/dilawar/epimargin in `tests` directory.

I tried to run the scripts available in `studies` folder but ran into few problems. Running a linter (`flake8`) exposed gaping issues that needs to be fixed (https://github.com/COVID-IWG/epimargin/issues/117). Authors also need to document the instructions to reproduce the images in studies folder accurately in the README files; or better still provide a pipeline e.g. `Makefile`  etc to automate the run.

Related issues: https://github.com/COVID-IWG/epimargin/issues/117

Community guidelines are missing: 1 is definitely missing. 2, and 3 is not stated explicitly in the repository but available at https://adaptivecontrol.org/. Probably authors want to add them to the repository as well.

- [x] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

Yes. My reading of the paper that the major intention behind the project is to quantify/access the impact of the pandemic and the trade-offs of various policies that govt might implement to tackle the pandemic such as lock-down etc.

- [x] A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?

Yes.

- [ ] State of the field: Do the authors describe how this software compares to other commonly-used packages?

There is no mention of any other software packages. I don't know if there are commonly-used packages.  I am aware of a few modeling studies in India itself e.g. https://cni-iisc.github.io/epidemic-simulator/docs/index.html but I am not sure if these are comparable. It would be nice to have a small table of comparisons with existing tools (if any).
