### 2021-07-23 5ec9529
update cwd() utility function to use __name__ and Path.cwd() (#118)
