### 2021-07-23 COVID-IWG/epimargin#116 @satejsoman
Hi @dilawar - thanks for flagging this. I've just merged #118, which uses the `__file__` builtin (when available) to construct a directory in the same location as the script. When that builtin is not available, we default to what `pathlib.Path.cwd()` provides.

We don't expect `/usr/bin` to be the default location for data; our intention was to create directories in the same folder as the client script. We do expect that client code calling this library will be run from a user-writable location, which I think is reasonable.

My testing indicates that `argv[0]` picks up the _script_ location rather than the binary location. As an example, consider this following test script, located in my home directory:
```python
from sys import argv
from pathlib import Path

print(argv[0])
print(Path(argv[0]).resolve().parent)
try:
    print(Path(__file__).resolve().parent)
except NameError:
    print("__file__ variable not available")
print(Path.cwd())
```

When I run `/usr/bin/python3 test.py`, I get:
```
test.py
/Users/satej
/Users/satej
/Users/satej
```

Does that make sense? Let me know if I've misunderstood something or have assumed something too specific about users' set ups.
