### 2021-07-23 COVID-IWG/epimargin#117 @satejsoman
Hi @dilawar, the scripts in the `studies` folder are not considered part of the core package; they are explorations and project-specific client code. I don't expect all the data scientists working with this package (some of whom have rolled off the project) to maintain these. I'd be happy to move these scripts to a separate `epimargin-studies` repository. At the very least, I'd request they be ignored in the code review.
