### 2021-07-23 01cc2ff
clarify community guidelines
