### 2021-07-23 6e5fd49
clarify language and dev steps
