+++ studies/vaccine_allocation/commons.py
@@ -13,7 +13,7 @@ from tqdm import tqdm
 """ Common data loading/cleaning functions and constants """
 
 data = Path("./data").resolve()
-# ext  = Path("/Volumes/dedomeno/covid/vax-nature").resolve()
+ext  = Path("/Volumes/dedomeno/covid/vax-nature").resolve()
 
 USD = 1/72
 
@@ -27,20 +27,19 @@ infectious_period = 1/gamma
 smooth = notched_smoothing(window)
 
 # simulation parameters
-# simulation_start = pd.Timestamp("April 15, 2021")
-simulation_start = pd.Timestamp("June 15, 2021")
+simulation_start = pd.Timestamp("April 15, 2021")
 
 num_sims = 1000
 focus_states = ["Tamil Nadu", "Punjab", "Maharashtra", "Bihar", "West Bengal"]
 # states to evaluate at state level (no district level data)
-coalesce_states = ["Delhi", "Manipur", "Dadra And Nagar Haveli And Daman And Diu", "Andaman And Nicobar Islands", "Telangana"]
+coalesce_states = ["Delhi", "Manipur", "Dadra And Nagar Haveli And Daman And Diu", "Andaman And Nicobar Islands"]
 
 # experiment_tag = "OD_IFR_Rtdownscale_fullstate"
 experiment_tag = "unitvaxhazard_TN_IFR"
 # epi_dst = tev_src = mkdir(ext/f"{experiment_tag}_epi_{num_sims}_{simulation_start.strftime('%b%d')}")
 # epi_dst = tev_src = mkdir(Path("/Volumes/dedomeno/covid/vax-nature/OD_IFR_Rtdownscale_fullstate_epi_1000_Apr15"))
-# epi_dst = tev_src = mkdir(Path("/Volumes/dedomeno/covid/vax-nature/all_india_coalesced_epi_1000_Apr15"))
-# tev_dst = fig_src = mkdir(ext/f"{experiment_tag}_tev_{num_sims}_{simulation_start.strftime('%b%d')}")
+epi_dst = tev_src = mkdir(Path("/Volumes/dedomeno/covid/vax-nature/all_india_coalesced_epi_1000_Apr15"))
+tev_dst = fig_src = mkdir(ext/f"{experiment_tag}_tev_{num_sims}_{simulation_start.strftime('%b%d')}")
 
 # misc
 survey_date = "October 23, 2020"
@@ -130,7 +129,7 @@ def get_state_timeseries(
     states = "*", 
     download: bool = False, 
     aggregation_cols = ["detected_state", "detected_district"], 
-    last_API_file: int = 30) -> pd.DataFrame:
+    last_API_file: int = 27) -> pd.DataFrame:
     """ load state- and district-level data, downloading source files if specified """
     paths = {"v3": [data_path(i) for i in (1, 2)], "v4": [data_path(i) for i in range(3, last_API_file)]}
     if download:
@@ -146,7 +145,7 @@ def get_state_timeseries(
             "Recovered":    "dR"
         })
 
-def case_death_timeseries(states = "*", download = False, aggregation_cols = ["detected_state", "detected_district"], last_API_file: int = 30):
+def case_death_timeseries(states = "*", download = False, aggregation_cols = ["detected_state", "detected_district"], last_API_file: int = 26):
     """ assemble a list of daily deaths and cases for consumption prediction """
     ts = get_state_timeseries(states, download)
     ts_index = pd.date_range(start = ts.index.get_level_values(-1).min(), end = ts.index.get_level_values(-1).max(), freq = "D")
[…]
