### 2021-07-23 d7d2a35
clarify community guidelines (#120)
* clarify community guidelines

* clarify language and dev steps

* Revert "clarify language and dev steps"

This reverts commit 6e5fd494f997d786e53bed5e849e3ce6427e405d.

* fix README after reversion
