### 2021-07-23 e15ee37
add comparison to state of the art
