+++ paper/paper.bib
@@ -84,3 +84,43 @@
   year={2021},
   institution={forthcoming}
 }
+
+@article{Palmer2021,
+  doi = {10.21105/joss.03290},
+  url = {https://doi.org/10.21105/joss.03290},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {63},
+  pages = {3290},
+  author = {Joseph Palmer and Katharine Sherratt and Richard Martin-Nielsen and Jonnie Bevan and Hamish Gibbs and Cmmid COVID-19 Working Group and Sebastian Funk and Sam Abbott},
+  title = {covidregionaldata: Subnational data for COVID-19 epidemiology},
+  journal = {Journal of Open Source Software}
+}
+
+@article{EpiEstim,
+    author = {Cori, Anne and Ferguson, Neil M. and Fraser, Christophe and Cauchemez, Simon},
+    title = "{A New Framework and Software to Estimate Time-Varying Reproduction Numbers During Epidemics}",
+    journal = {American Journal of Epidemiology},
+    volume = {178},
+    number = {9},
+    pages = {1505-1512},
+    year = {2013},
+    month = {09},
+    abstract = "{The quantification of transmissibility during epidemics is essential to designing and adjusting public health responses. Transmissibility can be measured by the reproduction number R, the average number of secondary cases caused by an infected individual. Several methods have been proposed to estimate R over the course of an epidemic; however, they are usually difficult to implement for people without a strong background in statistical modeling. Here, we present a ready-to-use tool for estimating R from incidence time series, which is implemented in popular software including Microsoft Excel (Microsoft Corporation, Redmond, Washington). This tool produces novel, statistically robust analytical estimates of R and incorporates uncertainty in the distribution of the serial interval (the time between the onset of symptoms in a primary case and the onset of symptoms in secondary cases). We applied the method to 5 historical outbreaks; the resulting estimates of R are consistent with those presented in the literature. This tool should help epidemiologists quantify temporal changes in the transmission intensity of future epidemics by using surveillance data.}",
+    issn = {0002-9262},
+    doi = {10.1093/aje/kwt133},
+    url = {https://doi.org/10.1093/aje/kwt133},
+    eprint = {https://academic.oup.com/aje/article-pdf/178/9/1505/17341195/kwt133.pdf},
+}
+
+@Article{ EpiNow,
+AUTHOR = { Abbott, S and Hellewell, J and Thompson, RN and Sherratt, K and Gibbs, HP and Bosse, NI and Munday, JD and Meakin, S and Doughty, EL and Chun, JY and Chan, YWD and Finger, F and Campbell, P and Endo, A and Pearson, CAB and Gimma, A and Russell, T and null, null and Flasche, S and Kucharski, AJ and Eggo, RM and Funk, S},
+TITLE = {Estimating the time-varying reproduction number of SARS-CoV-2 using national and subnational case counts [version 2; peer review: 1 approved with reservations]
+},
+JOURNAL = {Wellcome Open Research},
+VOLUME = {5},
+YEAR = {2020},
+NUMBER = {112},
+DOI = {10.12688/wellcomeopenres.16006.2}
+}
