### 2021-07-23 914830e
add comparison to state of the art (#121)
