+++ paper/paper.md
@@ -55,6 +55,8 @@ Included with the package are connectors and download utilities for common sourc
 
 This same modularity extends to both the estimation procedures and epidemiological models provided by `epimargin`. While the package includes a novel Bayesian estimator for key metrics, classical approaches based on rolling linear regressions and Markov chain Monte Carlo sampling are also included. The core model class in `epimargin` in which these estimates are used is known as a <i>compartmental</i> model: a modeled population is split into a number of mutually-exclusive compartments (uninfected, infected, recovered, vaccinated, etc) and flows between these compartments are estimated from empirical data. The exact choice of compartments and interactions is left to the modeler, but the package includes several commonly-used models, as well as variations customized for specific policy questions (such as large-scale migration during pandemics, or the effects of various vaccine distribution policies).
 
+For similar data downloading tools, see `covidregionaldata` [@Palmer2021]; for similar estimation tools, see `EpiEstim` [@EpiEstim] and `EpiNow2` [@EpiNow]. While many of these tools are used in conjunctionvwith each other, `epimargin` aims to offer tools for an end-to-end epidemiological workflow in one package, while offering the flexibility in estimator choice and data preparation methods.
+
 Attempts to use a compartmental model to drive policy decisions often treat the systems under study as deterministic and vary parameters such as the reproductive rate across a range deemed appropriate by the study authors [@bubar2021model]. This methodology complicates incorporation of recent disease data and the development of theories for why the reproductive rate changes due to socioeconomic factors external to the model. The incorporation of stochasticity into the models from the outset allows for the quantification of uncertainty and the illustration of a range of outcomes for a given public health policy under consideration.
 
 The `epimargin` package has been used to drive a number of research projects and inform policy decisions in a number of countries:
