### 2021-07-23 #3464 @satejsoman
Hi @dilawar, I've responded to/managed some of the issues you've brought up on their corresponding Github pages, and thank you for your thorough investigation. To elaborate on/respond to some of these points:

## Functionality
> Installation: Does installation proceed as outlined in the documentation?
> Yes. pip3 install apimargin succeeded on a Linux machine: it downloaded a platform-independent wheel. The installation from the source failed on my machine because of pinned version in requirements.txt. I think that requirements.txt file can be removed since setup.py files also enumerate the installation dependencies.

Generally, the `setup.py` and requirements.txt files have different uses. Python packaging practice indicates that `setup.py` have a minimal working set of requirement for end users to run out of the box, while `requirements.txt` contains the exact dependencies required to engage in development work on the package. (See this [stackoverflow post](https://stackoverflow.com/a/43659126) for an explanation of the guidelines we try to follow). For this reason, I would keep both `setup.py` and `requirements.txt`.

> After installation (from pypi), however, I ran into permission issues at the very first run.

We've switched to a more robust way of checking for the script directory location that should not try to write to `/usr/bin` based on our testing. See https://github.com/COVID-IWG/epimargin/issues/116 (closed by https://github.com/COVID-IWG/epimargin/pull/118).

## Documentation
>  Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
> A requirements.txt file is also given with pinned version of dependencies. User is expected use this file: pip3 install -r requiements.txt before pip3 install -e . but it did not succeed on Ubuntu/Debian and CentOS:8, For example, I did not find zip=4.1.0 (Invalid dependency version zipp==4.1.0 COVID-IWG/epimargin#115) anywhere in pypi.

We have fixed the typo in `requirements.txt` (see https://github.com/COVID-IWG/epimargin/pull/119)

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
> Yes. A simple example is given in the README file. I was able to reproduce the figure.
> And a lot of examples are available in studies folder. But instructions to run them are lacking. Ideally, each folder should have a pipeline based onMakefile, bash scripts etc.. Or authors should mention the steps to reproduce the results in README file in the folder itself. A few folders do not pass the basic lint tests (issue #117).

(quoting from the discussion in https://github.com/COVID-IWG/epimargin/issues/117): The scripts in the `studies/` folder are not considered part of the core package; they are explorations and project-specific client code. I don't expect all the data scientists working with this package (some of whom have rolled off the project) to maintain these. I'd be happy to move these scripts to a separate `epimargin-studies` repository. At the very least, I'd request they be ignored in the code review.

I am also unable to reproduce the failure of the `karnataka` script on Python 3.9 but given the exploratory nature of these scripts, I don't think this should be a blocker for acceptance. I will keep digging into this, but the core workflow does work on Python 3.9.

>  Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
> No. There are no automated tests. If there are some tests deep inside the codebase, I can't find a them via pytest etc. Authors should indicate somewhere in README.md if there is a test suite and how to run it. I've created one during my review https://github.com/dilawar/epimargin in tests directory.

Thank you for providing us with example tests - we are in the process of merging those into the main repository. I would argue the documentation provided in the tutorial counts as "manual steps described so that the functionality of the software can be verified".

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
> Community guidelines are missing: 1 is definitely missing. 2, and 3 is not stated explicitly in the repository but available at https://adaptivecontrol.org/. Probably authors want to add them to the repository as well.

We have added all three points to the README (see https://github.com/COVID-IWG/epimargin/pull/120)

> State of the field: Do the authors describe how this software compares to other commonly-used packages?
> There is no mention of any other software packages. I don't know if there are commonly-used packages. I am aware of a few modeling studies in India itself e.g. https://cni-iisc.github.io/epidemic-simulator/docs/index.html but I am not sure if these are comparable. It would be nice to have a small table of comparisons with existing tools (if any).

We have added discussions of packages that have functionality comparable to the submodules of `epimargin` (see https://github.com/COVID-IWG/epimargin/pull/121). We have discussed the advantages and design decisions of our package in the main paper but now explicitly add alternative packages that do similar things but take a more opinionated stance on the nature of the estimators provided.

----------
Let me know what the next steps are! Again, we are very thankful for your efforts and diligence.
