### 2021-07-23 COVID-IWG/epimargin#117 @dilawar
Hi @satejsoman . Thanks for the clarification. I think moving non-maintained scripts to a separate repo is a good idea since they don't work with the latest version.

It would be nice to have at least one non-trivial study that can be used as starting point and/or a demonstration of the capabilities.
