### 2021-07-24 COVID-IWG/epimargin#123 @dilawar
Lint errors in core package
https://github.com/openjournals/joss-reviews/issues/3464 | Related #117.

I am running a few lint tools e.g., `pylint` and `pyflakes` to locate errors in the module. This is related to #117 that also used `studies` folder. Since that may not be part of final release (https://github.com/COVID-IWG/epimargin/issues/117#issuecomment-885935277), this run restricts itself to the core files.

I am working on my fork and send a PR soon that will fix these issues if possible.

In this run, I am using poetry to manage the dependencies. Managing dependencies in Python has been less than ideal (`poetry` helps but the author need not add this to their DevOps).

```
$ poetry run pylint -E epimargin
[dilawars@chutki epimargin (poetry)]$ poetry run pylint -E epimargin
************* Module epimargin.policy
epimargin/policy.py:205:15: E1101: Instance of 'VaccinationPolicy' has no 'daily_doses' member (no-member)
************* Module epimargin.estimators
epimargin/estimators.py:25:18: E1101: Method 'params' has no 'join' member (no-member)
************* Module epimargin.plots
epimargin/plots.py:62:0: E0102: function already defined line 14 (function-redefined)
epimargin/plots.py:194:0: E0102: function already defined line 14 (function-redefined)
************* Module epimargin.etl.covid19india
epimargin/etl/covid19india.py:348:22: E1101: Instance of 'TextFileReader' has no 'set_index' member (no-member)
************* Module epimargin.etl.devdatalab
epimargin/etl/devdatalab.py:28:57: E1101: Instance of 'TextFileReader' has no 'O_StateCensus2011' member (no-member)
************* Module epimargin.etl.csse
epimargin/etl/csse.py:27:8: E1101: Instance of 'TextFileReader' has no 'assign' member (no-member)
************* Module epimargin.etl
epimargin/etl/__init__.py:2:11: E0604: Invalid object 'download_data' in __all__, must contain only strings (invalid-all-object)
[dilawars@chutki epimargin (poetry)]$\
```
