+++ studies/vaccine_allocation/commons.py
@@ -38,8 +38,8 @@ coalesce_states = ["Delhi", "Manipur", "Dadra And Nagar Haveli And Daman And Diu
 experiment_tag = "unitvaxhazard_TN_IFR"
 # epi_dst = tev_src = mkdir(ext/f"{experiment_tag}_epi_{num_sims}_{simulation_start.strftime('%b%d')}")
 # epi_dst = tev_src = mkdir(Path("/Volumes/dedomeno/covid/vax-nature/OD_IFR_Rtdownscale_fullstate_epi_1000_Apr15"))
-epi_dst = tev_src = mkdir(Path("/Volumes/dedomeno/covid/vax-nature/all_india_coalesced_epi_1000_Apr15"))
-tev_dst = fig_src = mkdir(ext/f"{experiment_tag}_tev_{num_sims}_{simulation_start.strftime('%b%d')}")
+# epi_dst = tev_src = mkdir(Path("/Volumes/dedomeno/covid/vax-nature/all_india_coalesced_epi_1000_Apr15"))
+# tev_dst = fig_src = mkdir(ext/f"{experiment_tag}_tev_{num_sims}_{simulation_start.strftime('%b%d')}")
 
 # misc
 survey_date = "October 23, 2020"
@@ -147,7 +147,7 @@ def get_state_timeseries(
 
 def case_death_timeseries(states = "*", download = False, aggregation_cols = ["detected_state", "detected_district"], last_API_file: int = 26):
     """ assemble a list of daily deaths and cases for consumption prediction """
-    ts = get_state_timeseries(states, download)
+    ts = get_state_timeseries(states, download, aggregation_cols, last_API_file)
     ts_index = pd.date_range(start = ts.index.get_level_values(-1).min(), end = ts.index.get_level_values(-1).max(), freq = "D")
 
     return ts.unstack(-1)\
