### 2021-08-05 e749524
resolve pylint issues
