+++ epimargin/etl/__init__.py
@@ -1,2 +1,2 @@
 from .commons import download_data
-__all__ = [download_data]
+__all__ = ["download_data"]
