+++ requirements.txt
@@ -4,6 +4,8 @@ astroid==2.4.1
 attrs==19.3.0
 backcall==0.1.0
 beautifulsoup4==4.9.3
+bleach==3.3.0
+bump2version==1.0.1
 bumpversion==0.6.0
 certifi==2020.6.20
 cftime==1.1.3
@@ -11,23 +13,27 @@ chardet==3.0.4
 click==7.1.2
 click-plugins==1.1.1
 cligj==0.5.0
+cloudpickle==1.6.0
 colorama==0.4.3
 commonmark==0.9.1
 cycler==0.10.0
 Cython==0.29.21
+dask==2021.4.1
 decorator==4.4.2
 descartes==1.1.0
+docutils==0.17.1
 fastprogress==0.2.3
 Fiona==1.8.17
 flake8==3.8.3
 Flask==1.1.2
 flat-table==1.1.1
+fsspec==2021.4.0
 GDAL==3.1.3
 geopandas==0.8.1
 geos==0.2.2
 h5py==2.10.0
 idna==2.9
-importlib-metadata==1.6.1
+importlib-metadata==4.5.0
 ipython==7.13.0
 ipython-genutils==0.2.0
 isort==4.3.21
@@ -35,24 +41,31 @@ itsdangerous==1.1.0
 jedi==0.17.0
 Jinja2==2.11.3
 joblib==0.14.1
+keyring==23.0.1
 kiwisolver==1.2.0
 lazy-object-proxy==1.4.3
 linearmodels==4.19
+locket==0.2.1
 lxml==4.6.3
+mapclassify==2.4.2
 MarkupSafe==1.1.1
 matplotlib==3.2.1
 mccabe==0.6.1
 munch==2.5.0
+mypy==0.910
 mypy-extensions==0.4.3
 netCDF4==1.5.3
+networkx==2.5.1
 numpy==1.18.2
 packaging==20.4
 pandas==1.0.3
 parso==0.7.0
+partd==1.2.0
 patsy==0.5.1
 pexpect==4.8.0
 pickleshare==0.7.5
 Pillow==8.2.0
+pkginfo==1.7.0
 pprintpp==0.4.0
 prompt-toolkit==3.0.5
 property-cached==1.6.4
@@ -68,7 +81,11 @@ pyproj==2.6.0
 pyreadr==0.4.0
 python-dateutil==2.8.1
 pytz==2019.3
+PyYAML==5.4.1
+readme-renderer==29.0
 requests==2.23.0
+requests-toolbelt==0.9.1
+rfc3986==1.5.0
 rich==2.3.0
 rope==0.17.0
 scikit-learn==0.22.2.post1
@@ -82,13 +99,16 @@ statsmodels==0.11.1
 Theano==1.0.5
 tikzplotlib==0.9.4
 toml==0.10.1
+toolz==0.11.1
 tqdm==4.45.0
 traitlets==4.3.3
+twine==3.4.1
 typed-ast==1.4.1
 typing-extensions==3.7.4.2
 urllib3==1.26.5
 urlpath==1.1.7
 wcwidth==0.1.9
+webencodings==0.5.1
 Werkzeug==1.0.1
 wrapt==1.12.1
 xarray==0.15.1
