### 2021-08-05 6f12a61
renames, refactors, and type annotations
