+++ epimargin/policy.py
@@ -1,20 +1,20 @@
 from abc import abstractmethod
 from itertools import product
-from typing import Dict, List, Optional, Tuple
+from typing import Dict, List, Optional, Set, Tuple
 
 import numpy as np
+from scipy.stats import multinomial as Multinomial
 from sklearn.metrics import auc
 
-from scipy.stats import multinomial as Multinomial
-
-from .models import SIR
+from .models import SIR, NetworkedSIR
 from .utils import weeks
+
 
 def AUC(curve):
     return auc(x = range(len(curve)), y = curve)
 
 # Adaptive Control policies
-def simulate_lockdown(model: SIR, lockdown_period: int, total_time: int, Rt0_mandatory: Dict[str, float], Rt0_voluntary: Dict[str, float], lockdown: np.matrix, migrations: np.matrix) -> SIR:
+def simulate_lockdown(model: NetworkedSIR, lockdown_period: int, total_time: int, Rt0_mandatory: Dict[str, float], Rt0_voluntary: Dict[str, float], lockdown: np.matrix, migrations: np.matrix) -> NetworkedSIR:
     """ simulate a lockdown where stringency is assumed to be reflected by Rt0_mandatory, with a return to Rt0_voluntary (and possible migration) once lockdown is over """
     return model.set_parameters(Rt0 = Rt0_mandatory)\
         .run(lockdown_period,  migrations = lockdown)\
@@ -22,7 +22,7 @@ def simulate_lockdown(model: SIR, lockdown_period: int, total_time: int, Rt0_man
         .run(total_time - lockdown_period, migrations = migrations)
 
 def simulate_adaptive_control(
-    model: SIR, 
+    model: NetworkedSIR, 
     initial_run: int, 
     total_time: int, 
     lockdown: np.matrix, 
@@ -31,7 +31,7 @@ def simulate_adaptive_control(
     beta_v: Dict[str, float], 
     beta_m: Dict[str, float], 
     evaluation_period: int = 2*weeks, 
-    adjacency: Optional[np.matrix] = None) -> SIR:
+    adjacency: Optional[np.matrix] = None) -> NetworkedSIR:
     """ simulate the Malani et al. adaptive lockdown policy where districts are assigned to lockdown stringency buckets based on Rt """
     n = len(model)
     model.set_parameters(Rt0 = R_m)\
@@ -40,7 +40,10 @@ def simulate_adaptive_control(
     gantt = []
     last_category = dict()
     while days_run < total_time:
-        Gs, Ys, Os, Rs = set(), set(), set(), set()
+        Gs: Set[int] = set()
+        Ys: Set[int] = set()
+        Os: Set[int] = set()
+        Rs: Set[int] = set()
         categories = dict(enumerate([Gs, Ys, Os, Rs]))
         category_transitions = {}
         for (i, unit) in enumerate(model):
@@ -88,10 +91,10 @@ def simulate_adaptive_control(
         model.run(evaluation_period, phased_migration)
         days_run += evaluation_period
 
-    model.gantt = gantt 
+    model.gantt = gantt # type: ignore
     return model 
 
-def simulate_adaptive_control_MHA(model: SIR, initial_run: int, total_time: int, lockdown: np.matrix, migrations: np.matrix, R_m: Dict[str, float], beta_v: Dict[str, float], beta_m: Dict[str, float], evaluation_period = 2*weeks):
+def simulate_adaptive_control_MHA(model: NetworkedSIR, initial_run: int, total_time: int, lockdown: np.matrix, migrations: np.matrix, R_m: Dict[str, float], beta_v: Dict[str, float], beta_m: Dict[str, float], evaluation_period = 2*weeks):
[…]
