### 2021-08-05 COVID-IWG/epimargin#130 @satejsoman
unify parallelizing mechanisms
Currently, `SIR` is parallelized by running separate models with different seeds, but `AgeSIRVD` instantiates a single model and advances a vector state through multivariate draws tied to a single seed.

todo:
[] Update  `SIR`  to match the latter.
[] Update plotting code and analysis functions to take in a single new `SIR` model rather than an array of models
