### 2021-08-05 4771b82
restructure class hierarchy to reflect different parallelizing mechanisms
