### 2021-08-05 0b0b88a
resolve linting issues  (#129)
* resolve pylint issues

* renames, refactors, and type annotations

* restructure class hierarchy to reflect different parallelizing mechanisms

* second round of linting checks
