+++ epimargin/models.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from typing import Dict, Iterator, Optional, Sequence, Tuple, Union
+from typing import Dict, Iterator, Optional, Sequence, Tuple, Union, List
 
 import geopandas as gpd
 import numpy as np
@@ -124,7 +124,7 @@ class SIR():
     # parallel poisson draws for infection
     def parallel_forward_epi_step(self, dB: int = 0, num_sims = 10000): 
         # get previous state 
-        S, I, R, D, N = (vector[-1].copy() for vector in (self.S, self.I, self.R, self.D, self.N))
+        S, I, R, D, N = (vector[-1] for vector in (self.S, self.I, self.R, self.D, self.N))
 
         # update state 
         Rt = self.Rt0 * S/N
@@ -148,9 +148,9 @@ class SIR():
 
         I -= (num_dead + num_recov)
 
-        S = S.clip(0)
-        I = I.clip(0)
-        D = D.clip(0)
+        S = max(0, S)
+        I = max(0, I)
+        D = max(0, D)
 
         N = S + I + R
         beta = (num_cases * N)/(b * S * I)
@@ -172,15 +172,15 @@ class SIR():
     # parallel binomial draws for infection
     def parallel_forward_binom_step(self, dB: int = 0, num_sims = 10000): 
         # get previous state 
-        S, I, R, D, N = (vector[-1].copy() for vector in (self.S, self.I, self.R, self.D, self.N))
+        S, I, R, D, N = (vector[-1] for vector in (self.S, self.I, self.R, self.D, self.N))
 
         # update state 
         Rt = self.Rt0 * S/N
         p = self.gamma * Rt * I/N
 
-        num_cases = binom.rvs(n = S.astype(int), p = p, size = num_sims)
-        self.upper_CI.append(binom.ppf(self.CI,     n = S.astype(int), p = p))
-        self.lower_CI.append(binom.ppf(1 - self.CI, n = S.astype(int), p = p))
+        num_cases = binom.rvs(n = S, p = p, size = num_sims)
+        self.upper_CI.append(binom.ppf(self.CI,     n = S, p = p))
+        self.lower_CI.append(binom.ppf(1 - self.CI, n = S, p = p))
 
         I += num_cases
         S -= num_cases
@@ -195,12 +195,11 @@ class SIR():
 
         I -= (num_dead + num_recov)
 
-        S = S.clip(0)
-        I = I.clip(0)
-        D = D.clip(0)
+        S = max(0, S)
+        I = max(0, I)
+        D = max(0, D)
 
         N = S + I + R
-        # beta = (num_cases * N)/(b * S * I)
 
         # update state vectors 
         self.Rt.append(Rt)
@@ -222,17 +221,20 @@ class SIR():
     def __repr__(self) -> str:
         return f"[{self.name}]"
 
-class Age_SIRVD(SIR):
-    """ age-structured compartmental model with a vaccinated class for each age bin """
+class Age_SIRVD():
+    """ age-structured compartmental model with a vaccinated class for each age bin 
+    note that the underlying parallelizing mechanism is different from that of SIR and NetworkedSIR
+
+    """
     def __init__(self,
         name:                str,           # name of unit
         population:          int,           # unit population
         dT0:        Optional[int]  = None,  # last change in cases, None -> Poisson random intro 
         Rt0:                 float = 1.9,   # initial reproductive rate,
-        S0:                  int   = 0,     # initial susceptible
-        I0:                  int   = 0,     # initial infected
-        R0:                  int   = 0,     # initial recovered
-        D0:                  int   = 0,     # initial dead
+        S0:                  np.array = np.array(0), # initial susceptibles
[…]
