### 2021-08-07 a8bb77c
Added a github actions file.
