+++ requirements.txt
@@ -114,4 +114,4 @@ wrapt==1.12.1
 xarray==0.15.1
 xlrd==1.2.0
 zipp==3.1.0
-semver==2.12.0
+semver==2.9.0
