### 2021-08-07 7897f97
Added scikit-learn required for core functionality
