### 2021-08-07 6fdfb27
Try to run one study after installing extra dependencie.
Test on Linux and MacOS

Its not straighforward to install geos on Windows. I'll write a separate
pipeline for it.
