+++ .github/workflows/python.yaml
@@ -9,7 +9,7 @@ jobs:
     strategy:
       matrix:
         python-version: [3.7, 3.9]
-        os: [ ubuntu-latest, windows-latest ]
+        os: [ ubuntu-latest, macos-latest ]
 
     steps:
       - uses: actions/checkout@v2
@@ -17,7 +17,15 @@ jobs:
         uses: actions/setup-python@v2
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Install and run example
+
+      - name: Install
         run: |
           python setup.py install
-          python studies/bihar/bihar_simulation.py
+
+      - name: install extra to run studies
+        run: |
+          python -m pip install tqdm
+
+      - name: Run tutorial
+        run: |
+          cd studies/tutorial/ && python tutorial.py
