### 2021-08-07 9e204cf
Squashed commit of the following:
commit 7ca2647bc381f6023c605b860c38534da5c780c4
Author: Dilawar Singh <dilawar.s.rajput@gmail.com>
Date:   Sat Aug 7 10:19:55 2021 +0530

    reverted toy_example.py

commit e1bba0eab1148014cf61dca7a9b62983fc79f5e4
Author: Dilawar Singh <dilawar.s.rajput@gmail.com>
Date:   Sat Aug 7 10:19:41 2021 +0530

    Added readthedocs files.
