+++ docs/index.rst
@@ -1,0 +1,22 @@
+Welcome to epimargin's documentation!
+=====================================
+
+.. image:: https://github.com/COVID-IWG/epimargin/raw/master/docs/logo.svg
+    :width: 300
+
+API Documentation
+==================
+
+.. automodule:: epimargin
+   :members:
+
+.. toctree::
+   :maxdepth: 3
+   :caption: Table of Contents
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
