### 2021-08-07 76caeb8
doc:rtd: run sphinx-apidoc before running final stage
