+++ docs/index.rst
@@ -7,12 +7,16 @@ Welcome to epimargin's documentation!
 API Documentation
 ==================
 
-.. automodule:: epimargin
-   :members:
+
+..
+    comment: module.rst and epimargin.rst are generated by sphinx-build at run time.
+    See conf.py file.
+
 
 .. toctree::
-   :maxdepth: 3
-   :caption: Table of Contents
+
+   modules
+
 
 Indices and tables
 ==================
