### 2021-08-07 0261117
hack: use environment variable to pass autodoc options to generator.
