### 2021-08-07 #3464 @dilawar
These are final points left on my checklist.

`tutorial.py` seems to work fine. There is one example embedded in the `README.md` file. I think these are enough for this item.

Additionally, I would appreciate if `docs/toy_example.py` can also be fixed or removed. And all examples are added to a test pineline. If you wish to use CI pipeline I sent over a PR, enable the test pipeline in CI. I will help with the integration.

I could not find API docs that are browsable. `pydoc3 epimargin` did work so I various IDE will be able to use it. It meets the minimal requirement.

Nonetheless, I added readthedocs based flow https://github.com/COVID-IWG/epimargin/pull/132 that may be useful for the users (this is optional).

- [ ] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I'd insist on adding `pytest` or any other unit testing framework to the development flow. Also, consider adding a CI pipeline (such as Github actions, Travis, gitlab-ci etc, Gihub actions can be cherry-picked from https://github.com/COVID-IWG/epimargin/pull/131).
