### 2021-08-08 #3464 @Nikoleta-v3
Thank you @dilawar for such a thorough review! I have gone over the issues and the discussions both here and  on the software repository.  I just wanted to say that regarding:

> Additionally, I would appreciate it if docs/toy_example.py can also be fixed or removed.

Agreed that: a user should be able to run all examples that are included in the repository without failures. This included the scripts under the folder `studies`.

> The scripts in the studies/ folder are not considered part of the core package; they are explorations and project-specific client code. I don't expect all the data scientists working with this package (some of whom have rolled off the project) to maintain these. I'd be happy to move these scripts to a separate epimargin-studies repository. At the very least, I'd request they be ignored in the code review.

👍🏻 Please do move them to a separate repository or keep them as they are and make sure they are maintained. Also, if they are moved to a new  repository and are not maintained please do state that in the readme.

This is mainly to help future new users  to the project that will be looking to understand how the package can help their work.  Thus,  is mainly to ensure that the package is documented to a satisfactory level which leads me to this:

> Nonetheless, I added readthedocs based flow COVID-IWG/epimargin#132 that may be useful for the users (this is optional).

Having a readthedocs is not a requirement for publication, and honestly https://epimargin.readthedocs.io/en/latest/epimargin.html  in its current form does not add much.  The current codebase is modular, the functions are documented and include type hints. These practices make the code documented well enough.

Nevertheless, as @dilawar demonstrated it is very easy to create a readthedocs page for the package. In the future you should look into changing the functions’ documentation style to numpy style (for example) so that the readthedocs “write” themselves https://numpy.org/doc/stable/docs/howto_document.html . If this is indeed  something that interests the project in the future I would also be very happy to help with a `PR`.

_ps if you are interested in “how to write documentation” I often recommend the  Diátaxis framework https://diataxis.fr which is what I also use in my projects._

Lastly,

> I'd insist on adding pytest or any other unit testing framework to the development flow. Also, consider adding a CI pipeline (such as Github actions, Travis, gitlab-ci etc, Gihub actions can be cherry-picked from COVID-IWG/epimargin#131).

Authors are strongly encouraged to include an automated test suite covering the core functionality of their software. However, `documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)`.

which I believe comes back to making sure that the scripts that are included in the repository run without failures and the output is expected. If I tried to run one of your  “example scripts” as a new user and it failed I would assume that something is broken in the codebase.

> And all examples are added to a test pineline. If you wish to use CI pipeline I sent over a PR, enable the test pipeline in CI. I will help with the integration

This is a great idea, but it’s again not required of you. However, @dilawar has already offered their help, and I would also be happy to assist if this is something you would be interested in learning/including in the project.

------

A `summary`:

The package is documented well enough as it is. Including scripts such as `toy_example.py` and others from `studies` can only further help the documentation of your project and they should be included.

To ensure that the package is “tested” the scripts you decide to include in repository should be able to run without failure and produce the expected output.

`Optional (possible future) points regarding`

**Documentation**:

- Tweak the functions documentation style and generate a readthedocs

**Testing**:

- plug in the example scripts to a CI
- add pytest or any other unit testing framework to the development flow.
