### 2021-08-10 d3721a1
remove studies from core repo
