+++ studies/indonesia/jakarta_district_simulation.py
@@ -1,147 +1,0 @@
-from logging import getLogger
-from pathlib import Path
-from typing import Dict, Optional, Sequence, Tuple
-
-import epimargin.plots as plt
-import geopandas as gpd
-import numpy as np
-import pandas as pd
-from epimargin.estimators import analytical_MPVS
-from epimargin.models import SIR, NetworkedSIR
-from epimargin.policy import simulate_adaptive_control, simulate_lockdown
-from epimargin.smoothing import convolution, notched_smoothing
-from epimargin.utils import days, setup, weeks
-from scipy.spatial import distance_matrix
-from tqdm import tqdm
-
-from jakarta_granular import replacements
-
-logger = getLogger("DKIJ.AC")
-
-dkij_drop_cols = [
-    'age', 'sex', 'fever', 'temp', 'cough', 'flu', 'sore_throat', 'shortness_breath', 'shivering', 'headache', 'malaise', 'muscle_pain',
-    'nausea_vomiting', 'abdominal_pain', 'diarrhoea', 'date_recovered',
-    'date_died', 'heart_disease', 'diabetes', 'pneumonia', 'hypertension', 'malignant',
-    'immunology_disorder', 'chronic_kidney', 'chronic_liver', 'copd',
-    'obesity', 'pregnant', 'tracing', 'otg', 'icu', 'intubation', 'ecmo',
-    'criteria_cases', 'age_group', 'age_group2', 'date_discharge',
-    'patient_status', 'death'
-]
-
-shp_drop_cols = ['GID_0', 'NAME_0', 'GID_1', 'NAME_1', 'NL_NAME_1', 'GID_2', 'VARNAME_2', 'NL_NAME_2', 'TYPE_2', 'ENGTYPE_2', 'CC_2', 'HASC_2']
-
-
-def model(districts, populations, cases, seed) -> NetworkedSIR:
-    units = [
-        SIR(district, populations[i], 
-        dT0 = cases[i], 
-        R0 = 0, 
-        D0 = 0, 
-        mobility = 0.000001)
-        for (i, district) in enumerate(districts)
-    ]
-    return NetworkedSIR(units, random_seed=seed)
-
-def run_policies(
-        district_cases:  Dict[str, pd.DataFrame], # timeseries for each district 
-        populations:     pd.Series,               # population for each district
-        districts:       Sequence[str],           # list of district names 
-        migrations:      np.matrix,               # O->D migration matrix, normalized
-        gamma:           float,                   # 1/infectious period 
-        Rmw:             Dict[str, float],        # mandatory regime R
-        Rvw:             Dict[str, float],        # voluntary regime R
-        lockdown_period: int,                     # how long to run lockdown 
-        total:           int   = 90*days,         # how long to run simulation
-        eval_period:     int   = 2*weeks,         # adaptive evaluation period
-        beta_scaling:    float = 1.0,             # robustness scaling: how much to shift empirical beta by 
-        seed:            int   = 0                # random seed for simulation
-    ):
-    lockdown_matrix = np.zeros(migrations.shape)
-
-    # lockdown 1
-    model_A = model(districts, populations, district_cases, seed)
-    simulate_lockdown(model_A, lockdown_period, total, Rmw, Rvw, lockdown_matrix, migrations)
-
-    # lockdown 1
-    model_B = model(districts, populations, district_cases, seed)
-    simulate_lockdown(model_B, lockdown_period + 2*weeks, total, Rmw, Rvw, lockdown_matrix, migrations)
-
-    # lockdown + adaptive controls
[…]
