### 2021-08-10 9722ab0
remove studies from core repo (#133)
