### 2021-08-10 bb4e6f0
update toy example to use new classes and package names
