### 2021-08-10 b5f1ac1
update toy example to use new classes and package names (#134)
