+++ docs/toy_example.py
@@ -1,9 +1,8 @@
+import matplotlib.pyplot as plt
 import numpy as np
+from epimargin.models import NetworkedSIR, SIR
+from epimargin.plots import plot_SIRD
 from scipy.spatial import distance_matrix
-import matplotlib.pyplot as plt 
-
-from adaptive.model import Model, ModelUnit
-from adaptive.plotting import plot_SIRD
 
 names  = [f"state{i}" for i in ( 1, 2, 3, 4)]
 pops   = [k * 1000    for k in (10, 1, 2, 2)]
@@ -16,9 +15,12 @@ P *= np.array(pops)[:, None]  # weight by destination population
 P /= P.sum(axis = 0)          # normalize 
 
 # states 
-units = lambda rr0: [ModelUnit(name = name, population = pop, RR0 = rr0) for (name, pop) in zip(names, pops)]
+units = lambda Rt0: [SIR(name = name, population = pop, Rt0 = Rt0) for (name, pop) in zip(names, pops)]
 
 # run and plot 
-m1_9 = Model(units(1.9), P, 0).run(200)
-plot_SIRD(m1_9, title = "Four-State Toy Model", xlabel="Time", ylabel="S/I/R/D", subtitle="No Adaptive Controls")
-plt.show()
+networked_model = NetworkedSIR(units(1.9), P, 0).run(200)
+plot_SIRD(networked_model)\
+    .title("Four-State Toy Model")\
+    .xlabel("Time")\
+    .ylabel("S/I/R/D")\
+    .show()
