+++ requirements.txt
@@ -91,6 +91,7 @@ rope==0.17.0
 scikit-learn==0.22.2.post1
 scipy==1.4.1
 seaborn==0.10.0
+semver==2.9.0
 Shapely==1.7.0
 six==1.14.0
 sklearn==0.0
