### 2021-08-10 COVID-IWG/epimargin#136 @satejsoman
add pytest suite
We need a more comprehensive unit testing suite as the functionality has matured.
