### 2021-08-10 COVID-IWG/epimargin#136 @satejsoman
to do:

- [ ] add unit tests
- [ ] add tests to CI pipeline
