### 2021-08-10 #3464 @satejsoman
Hi @dilawar and @Nikoleta-v3 - thanks again.

Here are the changes that have been implemented:
- the `studies` folder has been moved to [its own repository](https://github.com/COVID-IWG/epimargin-studies)  (see: https://github.com/COVID-IWG/epimargin/pull/133)
- `docs/toy_example.py` has been updated to refer to the latest classes/package names and runs out of the box (see: https://github.com/COVID-IWG/epimargin/pull/134)
- Github Actions have been set up to run the tutorial (we cherrypicked your template, @dilawar) (see: https://github.com/COVID-IWG/epimargin/pull/135)

---
Here are the future changes we are tracking:
- set  up a more comprehensive `pytest` suite (see: https://github.com/COVID-IWG/epimargin/issues/136)
- add the `pytest` suite to the CI pipeline (see: same issue as above)

We expect to onboard a developer in the next few weeks and will be able to proceed with these fairly shortly from there!
