### 2021-08-10 #3464 @Nikoleta-v3
👋🏻 thank you for your work @satejsoman

> Github Actions have been set up to run the tutorial (we cherrypicked your template, @dilawar) (see: cherrypick GH actions from review branch COVID-IWG/epimargin#135)

Fantastic! 🎉  Note that Github Actions haven't  run yet (https://github.com/COVID-IWG/epimargin/actions).

I believe it's because there hasn't been a "push" to the project since you setup the CI, and CI is only going to run after a "push". This is because of [line](https://github.com/COVID-IWG/epimargin/blob/c6dcefa6568091a74ce78da675a0311e19a82258/.github/workflows/python.yaml#L3).

If Github Actions run now you are going to get a failure because:

```
- name: Run tutorial
      run: |
        cd studies/tutorial/ && python tutorial.py
```

`studies/tutorial/ ` does not exist anymore.

You need to move `tutorial.py` back in the main repository.

I recommend the following:

1. Run Github Actions and let them fail ❌
2.  Move `studies/tutorial/tutorial.py` to `doc/tutorial.py`
3. Tweak README to say `doc/tutorial` instead of `studies/tutorial/ `.
4. Change `python.yaml` to run `doc/tutorial` and `doc/toy_example`. Tests should pass ✅
5. Add readme in `doc` to explain (very briefly) the usage of the files `Create install.md`, `tutorial.py`, `toy_example`.
6. Add a brief section on your main `readme` exampling what it's currently being tested with the CI.

Feel free to ping if something is unclear  _(probably my fault for explaining things badly)_ 👍🏻 😄
