### 2021-08-10 ed71667
add back in tutorial.py and wire up CI
