+++ docs/tutorial.py
@@ -1,0 +1,110 @@
+# preliminary set up
+from itertools import cycle
+
+import epimargin.plots as plt
+import numpy as np
+import pandas as pd
+from epimargin.utils import setup
+
+(data, figs) = setup()
+plt.set_theme("minimal")
+
+# download, load, and clean data
+from epimargin.etl import download_data
+from epimargin.smoothing import notched_smoothing
+
+# a snapshot of this csv is checked into the repo at data/tutorial_timeseries.csv in case you run into download problems
+download_data(data, "districts.csv", "https://api.covid19india.org/csv/latest/") 
+
+daily_reports = pd.read_csv(data / "districts.csv", parse_dates = ["Date"])\
+    .rename(str.lower, axis = 1)\
+    .set_index(["state", "district", "date"])\
+    .sort_index()\
+    .loc["Maharashtra", "Mumbai"]
+daily_cases = daily_reports["confirmed"]\
+    .diff()\
+    .clip(lower = 0)\
+    .dropna()\
+
+smoother = notched_smoothing(window = 5)
+smoothed_cases = pd.Series(
+    data  = smoother(daily_cases),
+    index = daily_cases.index
+)
+
+# plot raw and cleaned data 
+beg = "December 15, 2020"
+end = "March 1, 2021"
+training_cases = smoothed_cases[beg:end]
+
+plt.scatter(daily_cases[beg:end].index, daily_cases[beg:end].values, color = "black", s = 5, alpha = 0.5, label = "raw case count data")
+plt.plot(training_cases.index, training_cases.values, color = "black", linewidth = 2, label = "notch-filtered, smoothed case count data")
+plt.PlotDevice()\
+    .l_title("case timeseries for Mumbai")\
+    .axis_labels(x = "date", y = "daily cases")\
+    .legend()\
+    .adjust(bottom = 0.15, left = 0.15)\
+    .format_xaxis()\
+    .size(9.5, 6)\
+    .save(figs / "fig_1.svg")\
+    .show()
+
+# estimate Rt 
+from epimargin.estimators import analytical_MPVS
+
+(dates, Rt, Rt_CI_upper, Rt_CI_lower, *_) = analytical_MPVS(training_cases, smoother, infectious_period = 10, totals = False)
+plt.Rt(dates[1:], Rt[1:], Rt_CI_upper[1:], Rt_CI_lower[1:], 0.95, legend_loc = "upper left")\
+    .l_title("$R_t$ over time for Mumbai")\
+    .axis_labels(x = "date", y = "reproductive rate")\
+    .adjust(bottom = 0.15, left = 0.15)\
+    .size(9.5, 6)\
+    .save(figs / "fig_2.svg")\
+    .show()
+
+# set up model
+from epimargin.models import SIR
+
+num_sims = 100
+N0 = 12.48e6
+R0, D0 = daily_reports.loc[end][["recovered", "deceased"]]
+I0  = smoothed_cases[:end].sum()
+dT0 = smoothed_cases[end]
+S0  = N0 - I0 - R0 - D0
+Rt0 = Rt[-1] * N0 / S0
+no_lockdown = SIR(
+    name = "no lockdown", 
+    population = N0, 
+    dT0 = np.ones(num_sims) * dT0, Rt0 = np.ones(num_sims) * Rt0, I0 = np.ones(num_sims) * I0, R0 = np.ones(num_sims) * R0, D0 = np.ones(num_sims) * D0, S0 = np.ones(num_sims) * S0, infectious_period = 10
+)
+lockdown = SIR(
+    name = "partial lockdown", 
+    population = N0, 
[…]
