### 2021-08-10 f32e581
add back in tutorial.py and wire up CI (#137)
