+++ docs/tutorial.py
@@ -47,7 +47,7 @@ plt.PlotDevice()\
     .format_xaxis()\
     .size(9.5, 6)\
     .save(figs / "fig_1.svg")\
-    .show()
+    .show(block = False)
 
 # estimate Rt 
 from epimargin.estimators import analytical_MPVS
@@ -59,7 +59,7 @@ plt.Rt(dates[1:], Rt[1:], Rt_CI_upper[1:], Rt_CI_lower[1:], 0.95, legend_loc = "
     .adjust(bottom = 0.15, left = 0.15)\
     .size(9.5, 6)\
     .save(figs / "fig_2.svg")\
-    .show()
+    .show(block = False)
 
 # set up model
 from epimargin.models import SIR
@@ -107,4 +107,4 @@ plt.PlotDevice()\
     .size(9.5, 6)\
     .format_xaxis()\
     .save(figs / "fig_3.svg")\
-    .show()
+    .show(block = False)
