+++ .github/workflows/python.yaml
@@ -1,6 +1,6 @@
 name: Python package
 
-on: [push]
+on: [push, pull_request, workflow_dispatch]
 
 jobs:
   build:
