+++ .github/workflows/python.yaml
@@ -1,4 +1,4 @@
-name: Python package
+name: run tutorial
 
 on: [push, pull_request, workflow_dispatch]
 
