+++ .github/workflows/python.yaml
@@ -1,6 +1,12 @@
 name: run tutorial
 
-on: [push, pull_request, workflow_dispatch]
+on:
+  push:
+    branches: [master]
+  pull_request:
+    branches: [master]
+  workflow_dispatch: {}
+  
 
 jobs:
   build:
