+++ .github/workflows/python.yaml
@@ -19,6 +19,8 @@ jobs:
 
     steps:
       - uses: actions/checkout@v2
+          with:
+            ref: master
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v2
         with:
