+++ .github/workflows/python.yaml
@@ -1,22 +1,14 @@
 name: run tutorial
 
-on:
-  push:
-    branches: [master]
-  pull_request:
-    branches: [master]
-  workflow_dispatch: {}
-  
+on: [push, pull_request, workflow_dispatch]
 
 jobs:
   build:
-
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
         python-version: [3.7, 3.9]
         os: [ ubuntu-latest, macos-latest ]
-
     steps:
       - uses: actions/checkout@v2
         with:
@@ -25,15 +17,12 @@ jobs:
         uses: actions/setup-python@v2
         with:
           python-version: ${{ matrix.python-version }}
-
       - name: Install
         run: |
           python setup.py install
-
       - name: install extra to run studies
         run: |
           python -m pip install tqdm
-
       - name: Run tutorial
         run: |
           cd docs/tutorial/ && python tutorial.py
