+++ .github/workflows/python.yaml
@@ -11,8 +11,6 @@ jobs:
         os: [ ubuntu-latest, macos-latest ]
     steps:
       - uses: actions/checkout@v2
-        with:
-          ref: master
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v2
         with:
