### 2021-08-10 cfdeefa
Delete python-app.yml
