+++ .travis.yml
@@ -1,0 +1,15 @@
+dist: xenial 
+language: python
+python:
+  - "3.7"
+cache: pip
+before_install:
+  - sudo apt-get install libgdal-dev
+install:
+  - pip install wheel
+  - pip install -r requirements.txt
+  - pip install -e . 
+script: cd docs/tutorial/ && python tutorial.py
+branches:
+  only:
+  - master
