+++ .travis.yml
@@ -7,8 +7,7 @@ before_install:
   - sudo apt-get install libgdal-dev
 install:
   - pip install wheel
-  - pip install -r requirements.txt
-  - pip install -e . 
+  - python setup.py install
 script: cd docs/tutorial/ && python tutorial.py
 branches:
   only:
