+++ .travis.yml
@@ -7,7 +7,8 @@ before_install:
   - sudo apt-get install libgdal-dev
 install:
   - pip install wheel
-  - python setup.py install
+  - pip install -r requirements.txt
+  - pip install -e .
 script: cd docs/tutorial/ && python tutorial.py
 branches:
   only:
