### 2021-08-10 eb2805d
fix max call on numpy arrays
