### 2021-08-10 875adbd
remove mac os x from CI matrix
