+++ .github/workflows/tutorial-ci.yaml
@@ -8,7 +8,7 @@ jobs:
     strategy:
       matrix:
         python-version: [3.7, 3.9]
-        os: [ ubuntu-latest, macos-latest ]
+        os: [ ubuntu-latest ]
     steps:
       - uses: actions/checkout@v2
       - name: Set up Python ${{ matrix.python-version }}
