### 2021-08-10 #3464 @satejsoman
No worries @Nikoleta-v3, my fault for moving too quickly 😅

I've set up the Github Actions (and the [successful runs are showing up](https://github.com/COVID-IWG/epimargin/actions/workflows/tutorial-ci.yaml)) and have added the documentation to the README.

I elected to not add the toy_example to the CI since the more important workflows are captured in the tutorial.
