### 2021-08-11 #3464 @wxwx1993
> Hi @dilawar, [COVID-IWG/epimargin#129](https://github.com/COVID-IWG/epimargin/pull/129) has taken care of all the linting issues I could find with `pylint` and `mypy`. There is one temporary change to the class hierarchy that does not affect any user-facing code and it is being tracked here: [COVID-IWG/epimargin#130](https://github.com/COVID-IWG/epimargin/issues/130).
>
> With this merged, do you think remaining checklist items are ready to be marked as resolved? (i.e.: substantial scholarly effort, functionality, example usage, functionality documentation, "manual steps described so that the functionality of the software can be verified")
>
> @wxwx1993 - is there anything I can do to help with your review, like cutting a new release with the latest issues I've fixed and pushing it to PyPI?

Yes, that would be great!
