### 2021-08-11 #3464 @wxwx1993
### Software paper

I do not find the comparison to other commonly-used packages. I think CDC has the ensemble forecast model, which might be nice to compare with (https://www.cdc.gov/coronavirus/2019-ncov/science/forecasting/mathematical-modeling.html).
Also, I think it would be nice to also list the data input required for the software functions.
