### 2021-08-24 730d87b
fix argv reading in ipython
