### 2021-08-24 d845593
fix argv reading in ipython (#139)
* fix argv reading in ipython
