+++ .github/workflows/tutorial-ci.yaml
@@ -18,9 +18,6 @@ jobs:
       - name: Install
         run: |
           python setup.py install
-      - name: install extra to run studies
-        run: |
-          python -m pip install tqdm
       - name: Run tutorial
         run: |
-          cd docs/ && python tutorial.py
+          cd docs/ && python tutorial.py headless
