### 2021-08-24 eda80db
bump version to 0.6.1
