### 2021-08-24 fd8712a
add comments  explaining estimator arguments
