### 2021-08-24 #3464 @satejsoman
@wxwx1993 epimargin 0.6.1 has a [bugfix](https://github.com/COVID-IWG/epimargin/pull/139) that should allow the tutorial to work with `ipython -i`.

on the point of adding input data requirements to the manuscript, I prefer documenting the input timeseries and parameters of the included estimators in code (as done in the latest commit on master) rather than in the paper because the exact estimator inputs are methodological choices. For example, the [analytical mean-preserving variance spread estimator](https://github.com/COVID-IWG/epimargin/blob/master/epimargin/estimators.py#L44) can take either a timeseries of infections or deaths, and which is more appropriate is a modeling choice left to the client user.

The point we are trying to make in the paper itself is one of modularity: the end user can mix and match estimators and models as they see fit and there is nothing precluding them from using a completely different estimator (based on e.g. mobility patterns or virological rather population-level models) but plugging those estimates into a model provided out of the box. Does that make sense?
