### 2021-08-31 #3464 @satejsoman
Hi @wxwx1993, just checking in to see if you were able to run the tutorial on version 0.6.1?
