### 2021-08-31 #3464 @wxwx1993
@satejsoman Yes, I can successfully run the tutorial with iPython, but I am not able to run it in Jupiter Notebook. But I think it is fine.
