### 2021-08-31 #3464 @wxwx1993
I think your approach to document the input time-series and parameters of the included estimators in code is perfect for me. I do not have further major comments.
