### 2021-09-01 #3464 @dilawar
I have one unchecked item.

- [ ] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I guess this has been resolved by https://github.com/openjournals/joss-reviews/issues/3464#issuecomment-896383752 (also https://github.com/openjournals/joss-reviews/issues/3464#issuecomment-895890308)?
