### 2021-09-01 #3464 @dilawar
> I have one unchecked item.
>
>     * [ ]  Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
>
>
> I guess this has been resolved by [#3464 (comment)](https://github.com/openjournals/joss-reviews/issues/3464#issuecomment-896383752) (also [#3464 (comment)](https://github.com/openjournals/joss-reviews/issues/3464#issuecomment-895890308))?

- [x] CI is added: https://github.com/COVID-IWG/epimargin/actions/runs/1162050595 👍🏽 .
- [x] Outdated studies are not part of repo anymore.
- [x] Testing framework can be added in the future.

These closes the last item I had on my list.
