### 2021-09-02 #3464 @Nikoleta-v3
Thank you for your patience @satejsoman. The paper looks really good 👍🏻

Your package is a very useful tool, and it would be great to see it being used to analyse data across the globe. One question: Assume that I want to analyse data for Greece which have been downloaded from a different source and not `covid19india.org`. Would I have to tweak the format of my data to a "standard" one - so they can be loaded and analysed?

Two minor comments/questions:

- In the `readme` you refer to `studies/tutorial/data` which is not there anymore.
- This line is commented out with a question: https://github.com/COVID-IWG/epimargin/blob/fd8712ae8f2d2a6c785d046bd711f37bb3ec9d05/epimargin/estimators.py#L119. Is it meant to be here?
