### 2021-09-03 8a7a00c
added clarifying comments and removed unneeded ones (#140)
