### 2021-09-03 #3464 @satejsoman
Hi @Nikoleta-v3, the estimators generally work on anything that looks like a timeseries in pandas DataFrame form, with the assumption that the index is either an object interpretable as a timestamp, or a simple integer index (I've added clarifying comments to `epimargin/estimators.py`. Many data sources provide data in this format, and for sources that do not, it's usually a simple transformation (e.g. [our adapter for JHU's CSSE repository](https://github.com/COVID-IWG/epimargin/blob/master/epimargin/etl/csse.py#L35))

Thanks for bringing up that point! I've also fixed the references to `studies/tutorial/data` in the readme and have removed the unneeded comments.
