### 2021-09-04 #3464 @satejsoman
Hi @Nikoleta-v3, I've tagged and released version **0.6.2** of epimargin, and uploaded it to Zenodo and PyPI. The DOI is: 10.5281/zenodo.5430698
