+++ paper/paper.bib
@@ -39,7 +39,8 @@
   number={12},
   pages={e1008409},
   year={2020},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pcbi.1008409}
 }
 
 @article{bubar2021model,
