+++ requirements.txt
@@ -64,7 +64,7 @@ partd==1.2.0
 patsy==0.5.1
 pexpect==4.8.0
 pickleshare==0.7.5
-Pillow==8.2.0
+Pillow==8.3.2
 pkginfo==1.7.0
 pprintpp==0.4.0
 prompt-toolkit==3.0.5
